move gpu-related items in transport/simulation into code factory
